+++ lib/ooxml_parser/common_parser/common_data/alternate_content/drawing/graphic/shape/shape_properties/docx_shape_properties/text_box.rb
@@ -1,25 +1,22 @@
-# Class for working with TextBox (w:txbxContent)
+require_relative 'text_box/text_box_content'
+
 module OoxmlParser
-  class TextBox
-    # Parse TextBox List
-    # @param [Nokogiri::XML:Node] node with TextBox
-    # @return [Array] array of elements
-    def self.parse_list(node, parent: nil)
-      elements = []
-      text_box_content_node = node.xpath('w:txbxContent').first
-      unless text_box_content_node.nil?
-        text_box_content_node.xpath('*').each_with_index do |textbox_element, i|
-          case textbox_element.name
-          when 'p'
-            DocumentStructure.default_paragraph_style = DocxParagraph.new
-            DocumentStructure.default_paragraph_style.spacing = Spacing.new(0, 0.35, 1.15, :multiple)
-            elements << DocumentStructure.default_paragraph_style.dup.parse(textbox_element, i, parent: parent)
-          when 'tbl'
-            elements << Table.new(parent: parent).parse(textbox_element, i)
-          end
+  # Class for working with TextBox (v:textbox)
+  class TextBox < OOXMLDocumentObject
+    # @return [TextBoxContent] text box content
+    attr_reader :text_box_content
+
+    # Parse TextBox object
+    # @param node [Nokogiri::XML:Element] node to parse
+    # @return [TextBox] result of parsing
+    def parse(node)
+      node.xpath('*').each do |node_child|
+        case node_child.name
+        when 'txbxContent'
+          @text_box_content = TextBoxContent.new(parent: self).parse(node_child)
         end
       end
-      elements
+      self
     end
   end
 end